+++ app/components/UI/Bridge/BridgeSourceTokenSelector.tsx
@@ -1,9 +1,6 @@
 import React, { useCallback, useMemo } from 'react';
-import { StyleSheet } from 'react-native';
 import { useSelector, useDispatch } from 'react-redux';
 import { useNavigation } from '@react-navigation/native';
-import { useStyles } from '../../../component-library/hooks';
-import { Theme } from '../../../util/theme/models';
 import { TokenI } from '../Tokens/types';
 import { Hex } from '@metamask/utils';
 import { selectNetworkConfigurations } from '../../../selectors/networkController';
@@ -11,24 +8,11 @@ import { selectSelectedSourceChainIds, selectEnabledSourceChains, setSourceToken
 import { getNetworkImageSource } from '../../../util/networks';
 import { TokenSelectorItem } from './TokenSelectorItem';
 import { useSourceTokens, TokenIWithFiatAmount } from './useSourceTokens';
-import Button, { ButtonVariants } from '../../../component-library/components/Buttons/Button';
-import Routes from '../../../constants/navigation/Routes';
 import { useSortedSourceNetworks } from './useSortedSourceNetworks';
-import { MAX_NETWORK_ICONS, SourceNetworksButtonLabel } from './SourceNetworksButtonLabel';
+import { BridgeSourceNetworksBar, MAX_NETWORK_ICONS } from './BridgeSourceNetworksBar';
 import { BridgeTokenSelectorBase } from './BridgeTokenSelectorBase';
-import { IconName } from '../../../component-library/components/Icons/Icon';
-
-const createStyles = (params: { theme: Theme }) => {
-  const { theme } = params;
-  return StyleSheet.create({
-    networksButton: {
-      borderColor: theme.colors.border.muted,
-    },
-  });
-};
 
 export const BridgeSourceTokenSelector: React.FC = () => {
-  const { styles } = useStyles(createStyles, {});
   const dispatch = useDispatch();
   const navigation = useNavigation();
   const networkConfigurations = useSelector(selectNetworkConfigurations);
@@ -55,12 +39,6 @@ export const BridgeSourceTokenSelector: React.FC = () => {
     [handleTokenPress, networkConfigurations],
   );
 
-  const navigateToNetworkSelector = useCallback(() => {
-    navigation.navigate(Routes.MODAL.ROOT_MODAL_FLOW, {
-      screen: Routes.SHEET.BRIDGE_SOURCE_NETWORK_SELECTOR,
-    });
-  }, [navigation]);
-
   const networksToShow = useMemo(() =>
     sortedSourceNetworks
       .filter(({ chainId }) => selectedSourceChainIds.includes(chainId))
@@ -70,20 +48,14 @@ export const BridgeSourceTokenSelector: React.FC = () => {
 
   return (
     <BridgeTokenSelectorBase
-      renderNetworksBar={() => (
-        <Button
-          onPress={navigateToNetworkSelector}
-          variant={ButtonVariants.Secondary}
-          label={<SourceNetworksButtonLabel
-            networksToShow={networksToShow}
-            networkConfigurations={networkConfigurations}
-            selectedSourceChainIds={selectedSourceChainIds as Hex[]}
-            enabledSourceChains={enabledSourceChains}
-          />}
-          style={styles.networksButton}
-          endIconName={IconName.ArrowDown}
+      networksBar={
+        <BridgeSourceNetworksBar
+          networksToShow={networksToShow}
+          networkConfigurations={networkConfigurations}
+          selectedSourceChainIds={selectedSourceChainIds as Hex[]}
+          enabledSourceChains={enabledSourceChains}
         />
-      )}
+      }
       renderTokenItem={renderItem}
       tokensList={tokensList}
     />
